+++ src/routes/+page.svelte
@@ -8,30 +8,22 @@
 
   export let data: PageData;
 
-  const {
-    form,
-    errors,
-    enhance,
-    message,
-    delayed,
-    reset,
-    constraints,
-    fields
-  } = superForm(data.form, {
-    dataType: 'formdata',
-    async onUpdate({ form }) {
-      if (form.valid) {
-        await goto('?id=' + form.data.id);
+  const { form, errors, enhance, message, delayed, reset, constraints } =
+    superForm(data.form, {
+      dataType: 'formdata',
+      async onUpdate({ form }) {
+        if (form.valid) {
+          await goto('?id=' + form.data.id);
+        }
+      },
+      flashMessage: {
+        module: flashModule,
+        onError: (errorResult) => ({
+          type: 'error',
+          message: errorResult.error.message
+        })
       }
-    },
-    flashMessage: {
-      module: flashModule,
-      onError: (errorResult) => ({
-        type: 'error',
-        message: errorResult.error.message
-      })
-    }
-  });
+    });
 </script>
 
 <SuperDebug data={$form} />